add ducking, remove flicker, make clickable
Instead of hiding the statusdiv when hovered, just move it out of the way.
Delay 2ms before hiding to give time for other mouseover events to fire (resolves flicker).
Clicking the statusdiv now follows the link (in case it ever overlaps)
Rerranged CSS rules into an order that makes more sense to me.